+++ src/bin/lisp/welcome.lisp
@@ -1,11 +1,19 @@
 ; Function called automatically after image load.
 (fn istart ()
+  (load "user-post-image.lisp")
   (message "Ready."))
 
-; Have the universe list compressed.
-(= *universe (copy-list *universe*))
+; Load user file.
+(or (load "user-pre-image.lisp")
+    (message "INFO: You can create a file called \"user-pre-image.lisp\" to be loaded here."))
+
 (message "Cleaning up. Please wait...")
+; If compressed conses are enabled, have *UNIVERSE*
+; compressed better.
+(= *universe (copy-list *universe*))
 (print (gc))(out " bytes free.")(terpri)
+
+; Tell time it took to get here.
 (and (builtin? 'time)
      (number? +bps+)
   (block t
@@ -16,4 +24,9 @@
 
 (message "Saving 'image'...")
 (isave "image")
+
+; Load user image.
+(or (load "user-post-image.lisp")
+    (message "INFO: You can create a file called \"user-post-image.lisp\" to be loaded here and when the image is loaded on program start."))
+
 (message "Ready.")